organize file structure and moved existing tests into each page folder
