+++ client/src/pages/Routes.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import ProtectedRoute from '../components/Auth/hoc/ProtectedRoute';
-import Home from './Home/Home';
-import Auth from './Auth/Auth';
-import Events from './Events/Events';
-import Bookings from './Bookings/Bookings';
+import Home from './Home';
+import Auth from './Auth';
+import Events from './Events';
+import Bookings from './Bookings';
 
 import {
   ROOT_PATH,
