+++ client/src/tests/utils/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { AuthContextProvider } from '../../context/AuthContext';
+
 import {
   ApolloClient,
   InMemoryCache,
@@ -15,35 +16,32 @@ import { GRAPHQL_ENDPOINT } from '../../const';
 
 import '@testing-library/jest-dom';
 
-const httpLink = createHttpLink({
-  uri: `${process.env.REACT_APP_DEV_EVENT_BOOKER_API_URL}${GRAPHQL_ENDPOINT}`
-});
+// const httpLink = createHttpLink({
+//   uri: `${process.env.REACT_APP_DEV_EVENT_BOOKER_API_URL}${GRAPHQL_ENDPOINT}`
+// });
 
-const errorLink = onError(({ graphQLErrors, networkError }) => {
-  if (graphQLErrors)
-    graphQLErrors.forEach(({ message, locations, path }) =>
-      console.log(
-        `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`,
-      ),
-    );
+// const errorLink = onError(({ graphQLErrors, networkError }) => {
+//   if (graphQLErrors)
+//     graphQLErrors.forEach(({ message, locations, path }) =>
+//       console.log(
+//         `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`,
+//       ),
+//     );
 
-  if (networkError) console.log(`[Network error]: ${networkError}`);
-});
-const cache = new InMemoryCache();
-const client = new ApolloClient({
-  cache,
-  link: from([errorLink, httpLink]),
-});
+//   if (networkError) console.log(`[Network error]: ${networkError}`);
+// });
+// const cache = new InMemoryCache();
+// const client = new ApolloClient({
+//   cache,
+//   link: from([errorLink, httpLink]),
+// });
 
-// todo: remove apollo provider and use MockedProvider
 const Providers = ({ children }) => (
-  <ApolloProvider client={client}>
-    <BrowserRouter>
-      <AuthContextProvider>
-        {children}
-      </AuthContextProvider>
-    </BrowserRouter>
-  </ApolloProvider>
+  <BrowserRouter>
+    <AuthContextProvider>
+      {children}
+    </AuthContextProvider>
+  </BrowserRouter>
 );
 
 const customRender = (ui, options) => {
